Componentized Navbar and RandomFactCards. Added mockRandomFacts for dynamically rendering cards. Slightly adjusted layout to align cards nicely. Adjusted cards so that the content overflow will display a scrollbar correctly.
