+++ client/src/mocks/mockFacts.ts
@@ -1,0 +1,32 @@
+const facts = [
+  {
+    id: 0,
+    fact: 'Venetia Burney, who at age 11 suggested the name “Pluto” for the ninth planet in our solar system, lived to see it demoted to a dwarf planet in 2006.',
+  },
+  {
+    id: 1,
+    fact: 'Octopuses don’t have tentacles; they have arms. A tentacle only has suckers at its end, while a cephalopod arm has suckers for most of its length.',
+  },
+  {
+    id: 2,
+    fact: '30 Rock was known for its fast-paced writing. In 2014, one reporter calculated that there was an average of 7.44 jokes per minute.',
+  },
+  {
+    id: 3,
+    fact: 'According to a 2017 survey, "whatever" tops the list of words Americans find most annoying for the ninth year in a row.',
+  },
+  {
+    id: 4,
+    fact: 'A “contronym” is a word that’s its own opposite. For example: if you seed the lawn you add seeds, but if you seed a tomato you remove them.',
+  },
+  {
+    id: 5,
+    fact: 'The world’s oldest surviving piano is at the Metropolitan Museum of Art in New York City. The instrument dates back to 1720, and was made by an Italian man named Bartolomeo Cristofori?',
+  },
+  {
+    id: 6,
+    fact: 'Lake Hillier in Australia is nearly as salty as the Dead Sea. Scientists suspect that some salt-loving bacteria (chiefly Salinibacter ruber) and algae give the lake its pink hue.',
+  },
+];
+
+export { facts };
