+++ client/src/pages/Dashboard/Dashboard.tsx
@@ -3,21 +3,11 @@ import useStyles from './useStyles';
 import { useAuth } from '../../context/useAuthContext';
 import { useSocket } from '../../context/useSocketContext';
 import { useHistory } from 'react-router-dom';
-import {
-  AppBar,
-  Toolbar,
-  CircularProgress,
-  IconButton,
-  Typography,
-  Menu,
-  MenuItem,
-  CardContent,
-  Card,
-  Grid,
-  CardMedia,
-} from '@mui/material';
-import { Menu as MenuIcon, AccountCircle } from '@mui/icons-material';
+import { CircularProgress, Typography, Grid } from '@mui/material';
 import { Box } from '@mui/system';
+import { Navbar } from '../../components/Navbar';
+import { RandomFactCard } from '../../components/RandomFactCard';
+import { facts } from '../../mocks/mockFacts';
 
 export default function Dashboard(): JSX.Element {
   const classes = useStyles();
@@ -25,8 +15,6 @@ export default function Dashboard(): JSX.Element {
   const { loggedInUser } = useAuth();
   const { initSocket } = useSocket();
   const history = useHistory();
-
-  const [anchorEl, setAnchorEl] = React.useState<null | HTMLElement>(null);
 
   useEffect(() => {
     initSocket();
@@ -39,99 +27,20 @@ export default function Dashboard(): JSX.Element {
     return <CircularProgress />;
   }
 
-  const handleMenu = (event: React.MouseEvent<HTMLElement>) => {
-    setAnchorEl(event.currentTarget);
-  };
-
-  const handleClose = () => {
-    setAnchorEl(null);
-  };
-
   return (
     <>
-      <AppBar position="static">
-        <Toolbar>
-          <IconButton size="large" edge="start" color="inherit" aria-label="menu" sx={{ mr: 2 }}>
-            <MenuIcon />
-          </IconButton>
-          <Typography variant="h6" component="div" sx={{ flexGrow: 1 }}>
-            My App
-          </Typography>
-          {loggedInUser && (
-            <div>
-              <IconButton
-                size="large"
-                aria-label="account of current user"
-                aria-controls="menu-appbar"
-                aria-haspopup="true"
-                onClick={handleMenu}
-                color="inherit"
-              >
-                <AccountCircle />
-              </IconButton>
-              <Menu
-                id="menu-appbar"
-                anchorEl={anchorEl}
-                anchorOrigin={{
-                  vertical: 'bottom',
-                  horizontal: 'right',
-                }}
-                keepMounted
-                transformOrigin={{
-                  vertical: 'top',
-                  horizontal: 'right',
-                }}
-                open={Boolean(anchorEl)}
-                onClose={handleClose}
-              >
-                <MenuItem onClick={handleClose}>Profile</MenuItem>
-                <MenuItem onClick={handleClose}>My account</MenuItem>
-              </Menu>
-            </div>
-          )}
-        </Toolbar>
-      </AppBar>
-      <Grid sx={{ padding: 5 }} container spacing={2}>
-        <Grid item xs={12} justifyContent="center">
+      <Navbar />
+      <Grid sx={{ padding: 5 }} container rowSpacing={5} columnSpacing={2}>
+        <Grid item xs={12} sx={{ justifyContent: 'center' }}>
           <Typography variant="h4" component="h1">
             Welcome!
           </Typography>
         </Grid>
-        <Grid item xs={4}>
-          <Box sx={{ height: 100, bgcolor: '#26262B', borderRadius: '5px' }} />
-        </Grid>
-        <Grid item xs={4}>
-          <Box sx={{ height: 100, bgcolor: '#26262B', borderRadius: '5px' }} />
-        </Grid>
-        <Grid item xs={4}>
-          <Box sx={{ height: 100, bgcolor: '#26262B', borderRadius: '5px' }} />
-        </Grid>
-        <Grid item xs={4}>
-          <Box sx={{ height: 100, bgcolor: '#26262B', borderRadius: '5px' }} />
-        </Grid>
-        <Grid item xs={4}>
-          <Box sx={{ height: 100, bgcolor: '#26262B', borderRadius: '5px' }} />
-        </Grid>
-        <Grid item xs={4}>
-          <Box sx={{ height: 100, bgcolor: '#26262B', borderRadius: '5px' }} />
-        </Grid>
-        {/* <Card sx={{ flex: '1 0 auto' }}>
-            <CardMedia
-              component="img"
-              height="140"
-              image="https://picsum.photos/250/150"
-              alt="A randomly generated image"
-            />
-            <CardContent>
-              <Typography variant="h6" component="h6">
-                Did you know that,
-              </Typography>
-              <Typography variant="body1">
-                The world’s oldest surviving piano is at the Metropolitan Museum of Art in New York City. The instrument
-                dates back to 1720, and was made by an Italian man named Bartolomeo Cristofori?
-              </Typography>
-            </CardContent>
-          </Card> */}
+        {facts.map(({ id, fact }) => (
+          <Grid item key={id} xs={4} sx={{ display: 'flex', justifyContent: 'center', alignItems: 'center' }}>
+            <RandomFactCard fact={fact} />
+          </Grid>
+        ))}
       </Grid>
     </>
   );
